A bunch of uninteresting changes to make icc (and presumably other edgfe-based tools) whine less.
coastexp: fix bounds violation on monthString
delgpl: fix error text on fatal()
geoniche: tweak terminator
google: make seial an int, not long.
gpx: include xmlgeneric
lowrance: fix signedness writing waypoint count
mapsource make le_f(read|write)64 explictly void
nmea: add gratituous initialization
overlay: explicitly zero zome values
position: bloat 'nop' functions to eliminate need for cast.
shape: close->SHPClose.   misc warning fixes
tpg: include ctype
xmlgeneric.h: add many protos.

+++ trunk/gpsbabel/google.c
@@ -25,7 +25,7 @@ static char *script = NULL;
 
 FILE *fd;
 
-static long serial = 0;
+static int serial = 0;
 
 #define MYNAME "google"
 #define MY_CBUF 4096
